Edit posts.delete in PostController

+++ app/Http/Controllers/PostController.php
@@ -247,14 +247,18 @@ class PostController extends Controller
     public function destroy(post $post)
     {
         if (Auth::check()) {
-            $user = Auth::user();
+            if ($post->exists()) {
+                $user = Auth::user();
 
-            if ($user->id === $post->user_id) {
+                if ($user->id === $post->user_id) {
 
-                $post->delete();
-                return redirect(route('root'))->with('flash_message', '削除されました');
+                    $post->delete();
+                    return redirect(route('root'))->with('flash_message', '削除されました');
+                } else {
+                    return back()->with('flash_message', '投稿者でなければ削除できません');
+                }
             } else {
-                return back()->with('flash_message', '投稿者でなければ削除できません');
+                return redirect(route('root'))->with('flash_message', 'すでに存在しません');
             }
         } else {
             return back()->with('flash_message', '削除するにはログインする必要があります');
